+++ bindgen/src/bindings/cpp/gen_cpp/filters/mod.rs
@@ -16,6 +16,113 @@ use super::EnumStyle;
 
 type Result<T> = std::result::Result<T, askama::Error>;
 
+// Keywords and other commonly used words that are in fact macros that
+// expand to expressions (like `errno`).
+const RESERVED_CPP_WORDS: [&str; 100] = [
+    // Keywords
+    "alignas",
+    "alignof",
+    "and",
+    "and_eq",
+    "asm",
+    "atomic_cancel",
+    "atomic_commit",
+    "atomic_noexcept",
+    "auto",
+    "bitand",
+    "bitor",
+    "bool",
+    "break",
+    "case",
+    "catch",
+    "char",
+    "char8_t",
+    "char16_t",
+    "char32_t",
+    "class",
+    "compl",
+    "concept",
+    "const",
+    "consteval",
+    "constexpr",
+    "constinit",
+    "const_cast",
+    "continue",
+    "contract_assert",
+    "co_await",
+    "co_return",
+    "co_yield",
+    "decltype",
+    "default",
+    "delete",
+    "do",
+    "double",
+    "dynamic_cast",
+    "else",
+    "enum",
+    "explicit",
+    "export",
+    "extern",
+    "false",
+    "float",
+    "for",
+    "friend",
+    "goto",
+    "if",
+    "inline",
+    "int",
+    "long",
+    "mutable",
+    "namespace",
+    "new",
+    "noexcept",
+    "not",
+    "not_eq",
+    "nullptr",
+    "operator",
+    "or",
+    "or_eq",
+    "private",
+    "protected",
+    "public",
+    "reflexpr",
+    "register",
+    "reinterpret_cast",
+    "requires",
+    "return",
+    "short",
+    "signed",
+    "sizeof",
+    "static",
+    "static_assert",
+    "static_cast",
+    "struct",
+    "switch",
+    "synchronized",
+    "template",
+    "this",
+    "thread_local",
+    "throw",
+    "true",
+    "try",
+    "typedef",
+    "typeid",
+    "typename",
+    "union",
+    "unsigned",
+    "using",
+    "virtual",
+    "void",
+    "volatile",
+    "wchar_t",
+    "while",
+    "xor",
+    "xor_eq",
+    // Macros
+    "errno",
+    "assert",
+];
+
 #[derive(Clone)]
 pub(crate) struct CppCodeOracle;
 
@@ -40,7 +147,11 @@ impl CppCodeOracle {
     }
 
     pub(crate) fn var_name(&self, nm: &str) -> String {
-        nm.to_string().to_snake_case()
+        let mut name = nm.to_string().to_snake_case();
+        if RESERVED_CPP_WORDS.contains(&&*name) {
+            name.push('_');
+        }
+        name
     }
 
     pub(crate) fn object_names(&self, obj: &Object) -> (String, String) {