+++ fixtures/reserved-field-name/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    uniffi::generate_scaffolding("src/reserved_field_name.udl").unwrap();
+}